pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "spannery"
version = "0.2.0"
description = "A lightweight ORM for Google Cloud Spanner"
readme = "README.md"
requires-python = ">=3.10"
dependencies = ["google-cloud-spanner>=3.0.0"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[[project.authors]]
name = "Caio Pizzol"
email = "caiopizzol@gmail.com"

[project.license]
file = "LICENSE"

[tool.ruff]
line-length = 100  # Match your current Black config
target-version = "py310"  # Based on Spannery's requirements

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # Line too long (handled by formatter)
]

[tool.bandit]
exclude_dirs = [".tox", "venv", ".venv", "build", "dist"]
skips = ["B101"]  # Skip assert_used test

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--maxfail=2 -ra"
testpaths = ["tests"]

[tool.tox]
legacy_tox_ini = "[tox]\nenvlist = py310, py311, py312\nisolated_build = True\n\n[testenv]\ndeps =\n    -r ../requirements.txt\n    -r ../requirements-test.txt\n    python-dotenv\ncommands =\n    pytest ../tests --cov=../tests --disable-warnings -o log_cli=true --log-cli-level=INFO\n"
